requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2
deepseek==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
redis==5.0.1
prometheus-client==0.19.0
structlog==23.2.0
pyyaml==6.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
# psycopg2-binary==2.9.9  # Commented out to avoid PostgreSQL dependency
asyncpg==0.29.0
celery==5.3.4
flower==2.0.1
sentry-sdk[fastapi]==1.38.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
tiktoken==0.5.2 